+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Docstrings
 
 ### Changed
+- Made `CpxAp.add_module()` a private function as it should not be accessed by the user
 - Some minor code improvements
 - Added further explanation to example_cpxap_digital_output
 - Several backend changes in ap_module to clear lately ignored pylint issues